+++ tests/_Integration/Http/Html/SimpleListingTest.php
@@ -3,18 +3,27 @@
 namespace tests\_Integration\Http\Html;
 
 use Crwlr\Crawler\HttpCrawler;
+use Crwlr\Crawler\Loader\LoaderInterface;
 use Crwlr\Crawler\Steps\Html;
 use Crwlr\Crawler\Steps\Loading\Http;
 use Crwlr\Crawler\UserAgents\BotUserAgent;
 use Crwlr\Crawler\UserAgents\UserAgentInterface;
 
+use Psr\Log\LoggerInterface;
+
 use function tests\helper_generatorToArray;
+use function tests\helper_getFastLoader;
 
 it('gets all the links from a listing and gets data from the detail pages', function () {
     $crawler = new class () extends HttpCrawler {
         protected function userAgent(): UserAgentInterface
         {
             return new BotUserAgent('MyBot');
+        }
+
+        public function loader(UserAgentInterface $userAgent, LoggerInterface $logger): LoaderInterface
+        {
+            return helper_getFastLoader($userAgent, $logger);
         }
     };